Revert folder rename
