+++ sample/particles/meta.ts
