+++ sample/particles/probabilityMap.wgsl
